build: CI health -- aarch64 link, XE_AB opt-in, clang enum-range warnings
Three fixes for the state of the pipelines, each verified locally:

1. aarch64 still failed to link after JitHash.cpp was made portable,
   because portability is worthless for a TU that is never compiled: the
   NOT _M_X86 branch of pcsx2/CMakeLists.txt replaces pcsx2x86Sources
   wholesale and JitHash.cpp was not in the replacement list. It is now;
   its non-x86 arm is exactly the exported stub retro_deinit needs.
   Verified with aarch64-linux-gnu-g++ -fsyntax-only.

2. The A/B twin scaffolding is now OPT-IN (make XE_AB=1, cmake
   -DXE_AB=ON, define PCSX2_XE_AB). Default builds compile only the C89
   arm: the twin tokens are discarded by the preprocessor -- no template
   instantiation, no runtime branch, no counter. Compiling the twins into
   every build cost roughly +40% text and real compile time in the
   recompiler TUs on every platform for the benefit of exactly one
   verification workflow, and the macOS job started hitting its 1-hour
   runner ceiling under that weight; the workflow now pays for itself
   alone. The lean build boots and prints xe_site_hits 0 -- the counter
   line doubles as the config fingerprint. The XE_AB=1 build reproduces
   the full oracle: A/B identical across all seven hashes, 58,301 hits.
   Note for the Makefile: the flag must become a define next to
   PCSX2_C89_EMITTER's CXXFLAGS block; appending to COMMON_FLAGS before
   the include silently does nothing, which is the same trap the
   C89_EMITTER placement already documents from the other direction.
   Flag flips need a clean build -- object caching does not see CXXFLAGS.

3. Apple clang warns -Wtautological-constant-out-of-range-compare four
   times per emitter TU: the shim passes G1Type enums into E_REX_MEM's
   reg parameter, whose >=8 test the enum's value range can never
   satisfy. The comparison is semantically right -- group codes sit in
   the modrm reg field and never take REX.R -- so the fix is an int cast
   inside E_REX/E_REX_MEM, identical bytes, silent everywhere.

Also for the record, from reading the prepare-windows job that took
twenty minutes: its own section timestamps span sixteen seconds of work.
The remaining time elapsed on the runner before the first log line, and
the previous pipeline shows 2,788 seconds of queue on the same box --
the win10 shell runner was backlogged behind the long red Windows builds
this series fixes. Nothing in-repo executes during that window.

+++ common/emitter/c89emit.h
@@ -355,7 +355,10 @@ struct e_mem { int base; int index; int scale; e_sptr disp; };
         int rb_ = ((m).base  != E_NOREG && (m).base  >= 8) ? 1 : 0; \
         e_u8 rex_; \
         if (!E_NEEDS_SIB(m)) { rb_ = rx_; rx_ = 0; } \
-        rex_ = (e_u8)(0x40 | ((w)?8:0) | ((((reg)>=8)?1:0)<<2) | (rx_<<1) | rb_); \
+        /* (int) cast: callers pass group-code ENUMS here (they occupy the reg \
+ * field and never take REX.R); Apple clang's enum range analysis flags \
+ * the >=8 comparison as always-false and spams four warnings per TU. */ \
+        rex_ = (e_u8)(0x40 | ((w)?8:0) | ((((int)(reg)>=8)?1:0)<<2) | (rx_<<1) | rb_); \
         if (rex_ != 0x40) EW8((p), rex_); } while (0)
 
 /* EmitSibMagic for the register form. `after` counts bytes emitted after the
